Added error injection
added 2 alus with errors in them. To use them, just change in top the dut instantiation to have a 2 or 3 after the name of the module.

There is still the issue with the very first transaction sent, for some reason it has unknown when it reads the o_opcode, despite it being able to read properly in the other cases, so who knows why it wont work lol.